+++ src/Futhark/CodeGen/Backends/GenericC.hs
@@ -888,10 +888,11 @@ opaqueLibraryFunctions desc vds = do
   headerDecl (OpaqueDecl desc)
     [C.cedecl|int $id:free_opaque($ty:ctx_ty *ctx, $ty:opaque_type *obj);|]
 
-  return [C.cunit|
+  return
+    [C.cunit|
           int $id:free_opaque($ty:ctx_ty *ctx, $ty:opaque_type *obj) {
             int ret = 0, tmp;
-            $items:free_body
+            $items:(criticalSection free_body)
             free(obj);
             return ret;
           }